Let rqw customize the terminate signal
The default behavior of rqw is send SIGTERM, and then SIGKILL signals.
Although in Python, it's easier to catch SIGINT (2), because it's
automatically converted to a KeyboardInterput exception, and you don't
need to set up a specific signal handler to process it.

This commit adds an option "-signal" that allows to override the default
SIGTERM (15) with a different signal number.